cache altitude json too because it takes a while to compute the rasters. change the dashes to ampersands in the cache key because coordinates contain negative numbers

+++ server.js
@@ -152,6 +152,10 @@ app.get('/api/boundaries/:x1/:y1/:x2/:y2', function(request, response) {
 });
 
 app.get('/api/elevation-dump/:x1/:y1/:x2/:y2', function(request, response){
+  const cachedPath = path(__dirname + `/cache/elevation/elevation&${request.params.x1}&${request.params.y1}&${request.params.x2}&${request.params.y2}.json`);
+
+  if (fs.existsSync(cachedPath)) return response.json(JSON.parse(fs.readFileSync(cachedPath)));
+
   const query = `
     select to_json(ST_DumpValues(ST_Clip(ST_Union(rast),
       ST_MakeEnvelope(${request.params.x1}, ${request.params.y1}, ${request.params.x2}, ${request.params.y2}, 4326)
@@ -166,13 +170,15 @@ app.get('/api/elevation-dump/:x1/:y1/:x2/:y2', function(request, response){
       done();
       if (err) throw err;
       const vertices = result.rows[0].to_json.valarray
-      response.json({length: vertices.length, height: vertices[0].length, vertices: _.flatten(vertices)});
+      const json = {length: vertices.length, height: vertices[0].length, vertices: _.flatten(vertices)}
+      fs.writeFileSync(cachedPath, JSON.stringify(json));
+      response.json(json);
     });
   });
 });
 
 app.get('/api/terrain/:x/:y/:zoom', function(request, response){
-  const cachedImagePath = path(__dirname + `/cache/terrain/terrain${request.params.x}-${request.params.y}-${request.params.zoom}.jpg`);
+  const cachedImagePath = path(__dirname + `/cache/terrain/terrain&${request.params.x}&${request.params.y}&${request.params.zoom}.jpg`);
 
   if (fs.existsSync(cachedImagePath)) return response.end(fs.readFileSync(cachedImagePath), 'binary');
